Move auth implementaitons into a separate file.
